requirements.txt
datasets
transformers
torch
accelerate
flask
bitsandbytes
peft
logging
prometheus-flask-instrumentator
opentelemetry-sdk
opentelemetry-instrumentation-flask